Fix : Error DivideByZeroException
Previous to this commit, there was one case observed which was leading
to divide by zero exception as total permissions were equal to used
permissions. Such recommendations should not come up in the
recommendation list but were poping up. Due to which the code was
breaking.
This commit fix this error but making used_permissions =
total_permissions + 1. So with this logic _excess_permissions will never
be set to zero.